+++ tests/depth_prediction/test_dataset.py
@@ -7,13 +7,14 @@ import pytest
 from torch_kitti.depth_prediction import KittiDepthPredictionDataset
 
 
-def test_dataset_train(raw_sync_rect_path, depth_completion_path):
+@pytest.mark.parametrize("subset", ["train", "val"])
+def test_dataset_train(raw_sync_rect_path, depth_completion_path, subset):
 
-    # train subset
+    # train, val subset
     ds = KittiDepthPredictionDataset(
         raw_sync_rect_path,
         depth_completion_path,
-        subset="train",
+        subset=subset,
         load_stereo=True,
         load_previous=1,
     )
@@ -39,14 +40,49 @@ def test_dataset_train(raw_sync_rect_path, depth_completion_path):
         assert key in ex.keys()
 
 
-def test_dataset_value_error():
+def test_dataset_test(raw_sync_rect_path, depth_completion_path):
+
+    # test subset
+    with pytest.raises(ValueError):
+        ds = KittiDepthPredictionDataset(
+            raw_sync_rect_path,
+            depth_completion_path,
+            subset="test",
+            load_stereo=True,
+        )
+
+    with pytest.raises(ValueError):
+        ds = KittiDepthPredictionDataset(
+            raw_sync_rect_path,
+            depth_completion_path,
+            subset="test",
+            load_previous=1,
+        )
+
+    ds = KittiDepthPredictionDataset(
+        raw_sync_rect_path,
+        depth_completion_path,
+        subset="test",
+    )
+
+    assert len(ds) > 0
+
+    ex = ds[0]
+    keys = ["img", "gt", "intrinsics"]
+    for key in keys:
+        assert key in ex.keys()
+
+
+def test_dataset_value_error(raw_sync_rect_path, depth_completion_path):
 
     with pytest.raises(ValueError):
         KittiDepthPredictionDataset(
-            "..",
-            "..",
+            raw_sync_rect_path,
+            depth_completion_path,
             subset="something_wrong",
         )
 
     with pytest.raises(ValueError):
-        KittiDepthPredictionDataset("..", "..", load_previous=-23)
+        KittiDepthPredictionDataset(
+            raw_sync_rect_path, depth_completion_path, load_previous=-23
+        )